+++ Eikona/Tessa/ConnectorBundle/Resources/public/js/product/field/mam-field.js
@@ -85,6 +85,7 @@ define(
         return this.fieldTemplate({
           value,
           assets,
+          canEditAssetsInAkeneoUi: context.attribute.meta.canEditAssetsInAkeneoUi,
           isReadOnly: !this.isEditable()
         });
       },
@@ -217,8 +218,13 @@ define(
           prefix = 'PM-';
         }
 
+        const identifier = (this.productAttributes.model_type === 'product_model')
+          ? this.context.entity.code
+          : this.context.entity.identifier;
+
         const data = {
           ProductId: prefix + this.productAttributes.id,
+          identifier,
           attribute: JSON.stringify({
             code: this.attribute.code,
             type: this.attribute.type,