+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.2.0 (2020-04-09)
+* [feature] Automatically sync reference entities with TESSA
+* [feature] Added option to disable editing of assets in the Akeneo UI (editing then only works within the TESSA dialog)
+
 ## 1.1.1 (2020-02-04)
 * [bugfix] Allowed attributes of type 'pim_catalog_asset_collection' as main image
 * [bugfix] Fixed api for reference entities when using TESSA attributes